# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Chronik API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8788
    description: Development server
paths:
  /v1/ingest:
    post:
      summary: Universal ingest endpoint
      operationId: ingestEvent
      description: |
        Accepts JSON payloads either as a single object, an array of objects, or newline-delimited JSON.
        The target domain can be supplied via the `domain` query parameter (canonical) or as the `domain` field on the first
        object in the payload (legacy behavior, normalized by Chronik).

        **Canonical Endpoint**: `POST /v1/ingest?domain=heimgeist`
        **Legacy Alias**: `POST /ingest/heimgeist`

        **Limits**:
        - Max payload size: 1MB (configurable via `CHRONIK_MAX_BODY`).
        - Rate limit: 60 requests per minute (configurable via `CHRONIK_RATE_LIMIT`).

        **Validation**:
        - For domain `heimgeist`, strict validation applies. Required fields: `kind`, `version`, `id`, `meta` (with `occurred_at`, `role`), `data`.
      parameters:
        - name: domain
          in: query
          description: Optional target domain for the event. If omitted, the first payload object must contain `domain`.
          required: false
          schema:
            type: string
      security:
        - XAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Event'
                - type: array
                  items:
                    $ref: '#/components/schemas/Event'
            example:
              kind: "heimgeist.insight"
              version: 1
              id: "evt-12345678"
              meta:
                occurred_at: "2023-10-27T10:00:00Z"
                role: "agent"
              data:
                key: "value"
          application/x-ndjson:
            schema:
              type: string
              description: Newline-delimited JSON objects. Empty lines are ignored.
      responses:
        '202':
          description: Accepted. Body contains plain text `ok`.
          content:
            text/plain:
              schema:
                type: string
                example: ok
        '400':
          description: Bad request (invalid domain, JSON, payload, or missing domain information).
        '401':
          description: Unauthorized (missing or invalid `X-Auth` header).
        '411':
          description: Content-Length header missing.
        '413':
          description: Payload too large.
        '415':
          description: Unsupported content-type.
        '422':
          description: Semantic validation failed (e.g. summary too long).
        '429':
          description: Too many requests (rate limited). Clients should respect the `Retry-After` header.
        '507':
          description: Insufficient storage.
  /ingest/{domain}:
    post:
      deprecated: true
      summary: Deprecated ingest endpoint
      operationId: ingestEventDeprecated
      description: Replaced by **POST /v1/ingest**.
      parameters:
        - in: path
          name: domain
          required: true
          schema:
            type: string
      security:
        - XAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Event'
                - type: array
                  items:
                    $ref: '#/components/schemas/Event'
      responses:
        '202':
          description: Accepted (deprecated). Body contains plain text `ok`.
          content:
            text/plain:
              schema:
                type: string
                example: ok
        '400':
          description: Bad request (invalid domain, JSON, or payload).
        '401':
          description: Unauthorized (missing or invalid `X-Auth` header).
        '411':
          description: Content-Length header missing.
        '413':
          description: Payload too large.
        '429':
          description: Too many requests (rate limited). Clients should respect the `Retry-After` header.
        '507':
          description: Insufficient storage.
  /health:
    get:
      summary: Health check
      operationId: getHealth
      description: "Returns `{\"status\": \"ok\"}` when the service is ready."
      security:
        - XAuth: []
      responses:
        '200':
          description: Service healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '401':
          description: Unauthorized (missing or invalid `X-Auth` header).
  /version:
    get:
      summary: Service version
      operationId: getVersion
      description: Returns the current Chronik version string.
      security:
        - XAuth: []
      responses:
        '200':
          description: Version information
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string
components:
  securitySchemes:
    XAuth:
      type: apiKey
      in: header
      name: X-Auth
  schemas:
    Event:
      type: object
      description: |
        Generic event payload. Additional fields are preserved.
        No fields are unconditionally required. The `domain` field is required in the payload only
        when not supplied via the `domain` query parameter on /v1/ingest.
      required: []
      properties:
        domain:
          type: string
          description: Domain associated with the event. Required unless supplied via query parameter on /v1/ingest.
        summary:
          type: string
          description: Optional summary (maximum 500 characters).
          maxLength: 500
      additionalProperties: true